+++ lib/test_prof/recipes/rspec/let_it_be.rb
@@ -7,6 +7,12 @@ module TestProf
   # Just like `let`, but persist the result for the whole group.
   # NOTE: Experimental and magical, for more control use `before_all`.
   module LetItBe
+    Modifier = Struct.new(:scope, :block) do
+      def call(record, config)
+        block.call(record, config)
+      end
+    end
+
     class Configuration
       # Define an alias for `let_it_be` with the predefined options:
       #
@@ -17,10 +23,13 @@ module TestProf
         LetItBe.define_let_it_be_alias(name, **default_args)
       end
 
-      def register_modifier(key, &block)
+      # Register modifier by providing the name of key,
+      # optional scope (when to apply the modifier, on initialization (:initialize)
+      # or when accessed # via let (:let))
+      def register_modifier(key, on: :let, &block)
         raise ArgumentError, "Modifier #{key} is already defined for let_it_be" if LetItBe.modifiers.key?(key)
 
-        LetItBe.modifiers[key] = block
+        LetItBe.modifiers[key] = Modifier.new(on, block)
       end
 
       def default_modifiers
@@ -41,17 +50,18 @@ module TestProf
         @modifiers ||= {}
       end
 
-      def wrap_with_modifiers(mods, &block)
+      def wrap_with_modifiers(mods, on: :let, &block)
+        mods = mods.select { |k, val| LetItBe.modifiers.fetch(k).scope == on }
         return block if mods.empty?
 
         validate_modifiers! mods
 
-        -> {
+        proc do
           record = instance_eval(&block)
           mods.inject(record) do |rec, (k, v)|
             LetItBe.modifiers.fetch(k).call(rec, v)
           end
-        }
+        end
       end
 
       def module_for(group)
@@ -97,9 +107,11 @@ module TestProf
 
       options = default_options.merge(options)
 
+      initializer = LetItBe.wrap_with_modifiers(options, on: :initialize, &initializer)
+
       before_all(&initializer)
 
-      let_accessor = LetItBe.wrap_with_modifiers(options) do
+      let_accessor = LetItBe.wrap_with_modifiers(options, on: :let) do
         instance_variable_get(:"#{PREFIX}#{identifier}")
       end
 
@@ -213,7 +225,7 @@ if defined?(::ActiveRecord::Base)
       record
     end
 
-    config.register_modifier :freeze do |record, val|
+    config.register_modifier :freeze, on: :initialize do |record, val|
       if val == false
         TestProf::LetItBe::Freezer::Stoplist.stop!(record)
         next record